Task Sponsored polls and bug fixing

See merge request Antourage/widget.android!20

+++ ant-viewver/src/main/java/com/antourage/weaverlib/screens/list/rv/StreamItemDiffCallback.kt
@@ -42,7 +42,7 @@ internal class StreamItemDiffCallback(
         val diffBundle = Bundle()
 
         if (newStream.viewersCount != oldStream.viewersCount) {
-            newStream.viewersCount?.let { diffBundle.putInt(REFRESH_LIVE, it) }
+            newStream.viewersCount?.let { diffBundle.putLong(REFRESH_LIVE, it) }
         }
 
         if (newStream.viewsCount != oldStream.viewsCount) {
